pyproject.toml
[project]
name = "galmart"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.5.3",
    "coverage>=7.10.0",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "djoser>=2.3.3",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "markdown>=3.8.2",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.13.2",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
